\name{is_pinterest_url}
\alias{is_pinterest_url}
\title{Test whether a URL is supported}
\usage{is_pinterest_url(url)}
\arguments{\item{url}{A single character string.}}
\value{\code{TRUE} for a supported URL, otherwise \code{FALSE}.}
\description{Checks a Pinterest URL without making a network request.}
